pyproject.toml
[project]
name = "textent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.1.8",
    "kraken>=5.3.0",
    "lxml>=5.4.0",
    "pdf2image>=1.17.0",
    "pycases>=0.1.3",
    "pymupdf>=1.25.5",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "unidecode>=1.3.8",
    "yaltai>=2.0.5",
]
